test(smoke): retry smoke tests three times
we are often seeing intermittent failures in the smoke tests because the smoke test action is not ready yet or times out. This is a simple fix that will retry each request up to three times, to increase the chance of a 200 so that we don't have to wait another 5 minutes to run the entire test suite just because one request failed.